EZP-30544: Decorate fos http dispatcher instead of overriding
