+++ src/EzSystemsPlatformHttpCacheBundle.php
@@ -7,7 +7,7 @@
 namespace EzSystems\PlatformHttpCacheBundle;
 
 use EzSystems\PlatformHttpCacheBundle\DependencyInjection\ConfigResolver\HttpCacheConfigParser;
-use EzSystems\PlatformHttpCacheBundle\DependencyInjection\Compiler\HttpCachePass;
+use EzSystems\PlatformHttpCacheBundle\DependencyInjection\Compiler\VarnishCachePass;
 use EzSystems\PlatformHttpCacheBundle\DependencyInjection\Compiler\ResponseTaggersPass;
 use EzSystems\PlatformHttpCacheBundle\DependencyInjection\Compiler\KernelPass;
 use EzSystems\PlatformHttpCacheBundle\DependencyInjection\Compiler\DriverPass;
@@ -24,7 +24,7 @@ class EzSystemsPlatformHttpCacheBundle extends Bundle
         $container->addCompilerPass(new ResponseTaggersPass());
         $container->addCompilerPass(new KernelPass());
         $container->addCompilerPass(new DriverPass());
-        $container->addCompilerPass(new HttpCachePass());
+        $container->addCompilerPass(new VarnishCachePass());
 
         $this->registerConfigParser($container);
     }
